+++ x/gov/keeper/msg_server_test.go
@@ -41,6 +41,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					strings.Repeat("1", 300),
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr:    true,
@@ -55,6 +56,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr:    true,
@@ -69,6 +71,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr:    true,
@@ -83,6 +86,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr:    true,
@@ -97,6 +101,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr: false,
@@ -110,6 +115,7 @@ func (suite *KeeperTestSuite) TestSubmitProposalReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 			},
 			expErr: false,
@@ -147,8 +153,9 @@ func (suite *KeeperTestSuite) TestCancelProposalReq() {
 	msg, err := v1.NewMsgSubmitProposal(
 		[]sdk.Msg{bankMsg},
 		coins,
-		proposer.String(), "",
-		"title", "summary",
+		proposer.String(),
+		"", "title", "summary",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -221,6 +228,7 @@ func (suite *KeeperTestSuite) TestVoteReq() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -246,6 +254,7 @@ func (suite *KeeperTestSuite) TestVoteReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -289,6 +298,7 @@ func (suite *KeeperTestSuite) TestVoteReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -340,6 +350,7 @@ func (suite *KeeperTestSuite) TestVoteWeightedReq() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -366,6 +377,7 @@ func (suite *KeeperTestSuite) TestVoteWeightedReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -409,6 +421,7 @@ func (suite *KeeperTestSuite) TestVoteWeightedReq() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -459,6 +472,7 @@ func (suite *KeeperTestSuite) TestDepositReq() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -579,6 +593,7 @@ func (suite *KeeperTestSuite) TestLegacyMsgVote() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -604,6 +619,7 @@ func (suite *KeeperTestSuite) TestLegacyMsgVote() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -637,6 +653,7 @@ func (suite *KeeperTestSuite) TestLegacyMsgVote() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -688,6 +705,7 @@ func (suite *KeeperTestSuite) TestLegacyVoteWeighted() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -714,6 +732,7 @@ func (suite *KeeperTestSuite) TestLegacyVoteWeighted() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -747,6 +766,7 @@ func (suite *KeeperTestSuite) TestLegacyVoteWeighted() {
 					"",
 					"Proposal",
 					"description of proposal",
+					false,
 				)
 				suite.Require().NoError(err)
 
@@ -797,6 +817,7 @@ func (suite *KeeperTestSuite) TestLegacyMsgDeposit() {
 		"",
 		"Proposal",
 		"description of proposal",
+		false,
 	)
 	suite.Require().NoError(err)
 
@@ -1175,7 +1196,7 @@ func (suite *KeeperTestSuite) TestSubmitProposal_InitialDeposit() {
 			params.MinInitialDepositRatio = tc.minInitialDepositRatio.String()
 			govKeeper.SetParams(ctx, params)
 
-			msg, err := v1.NewMsgSubmitProposal(TestProposal, tc.initialDeposit, address.String(), "test", "Proposal", "description of proposal")
+			msg, err := v1.NewMsgSubmitProposal(TestProposal, tc.initialDeposit, address.String(), "test", "Proposal", "description of proposal", false)
 			suite.Require().NoError(err)
 
 			// System under test